Rename Settings to Connector settings
