+++ x-pack/plugins/enterprise_search/public/applications/enterprise_search_content/components/connector_settings/index.ts
@@ -5,4 +5,4 @@
  * 2.0.
  */
 
-export { Settings } from './settings';
+export { ConnectorSettings } from './connector_settings';
