+++ x-pack/plugins/enterprise_search/public/applications/enterprise_search_content/index.tsx
@@ -17,11 +17,16 @@ import { HttpLogic } from '../shared/http';
 import { KibanaLogic } from '../shared/kibana';
 import { VersionMismatchPage } from '../shared/version_mismatch';
 
+import { ConnectorSettings } from './components/connector_settings';
 import { ErrorConnecting } from './components/error_connecting';
 import { NotFound } from './components/not_found';
 import { SearchIndicesRouter } from './components/search_indices';
-import { Settings } from './components/settings';
-import { SETUP_GUIDE_PATH, ROOT_PATH, SEARCH_INDICES_PATH, SETTINGS_PATH } from './routes';
+import {
+  SETUP_GUIDE_PATH,
+  ROOT_PATH,
+  SEARCH_INDICES_PATH,
+  CONNECTOR_SETTINGS_PATH,
+} from './routes';
 
 export const EnterpriseSearchContent: React.FC<InitialAppData> = (props) => {
   const { config } = useValues(KibanaLogic);
@@ -71,8 +76,8 @@ export const EnterpriseSearchContentConfigured: React.FC<Required<InitialAppData
       <Route path={SEARCH_INDICES_PATH}>
         <SearchIndicesRouter />
       </Route>
-      <Route path={SETTINGS_PATH}>
-        <Settings />
+      <Route path={CONNECTOR_SETTINGS_PATH}>
+        <ConnectorSettings />
       </Route>
       <Route>
         <NotFound />
