+++ x-pack/plugins/enterprise_search/public/applications/enterprise_search_content/routes.ts
@@ -10,7 +10,7 @@ export const ROOT_PATH = '/';
 export const SETUP_GUIDE_PATH = '/setup_guide';
 
 export const SEARCH_INDICES_PATH = `${ROOT_PATH}search_indices`;
-export const SETTINGS_PATH = `${ROOT_PATH}settings`;
+export const CONNECTOR_SETTINGS_PATH = `${ROOT_PATH}connector_settings`;
 
 export const NEW_INDEX_PATH = `${SEARCH_INDICES_PATH}/new_index`;
 export const NEW_API_PATH = `${NEW_INDEX_PATH}/api`;
